AppWorld Integration for rLLM (#235)
* Add skeleton for AppWorldEnv and AppWorldReactAgent

This commit is the initial code change that introduces AppWorld integration with rLLM. It only adds the skeleton for the AppWorldEnv and AppWorldReactAgent classes that inherit the base classes.

* Basic implmentations and annotations of AppWorldEnv

This commit implements the basic steps of init, rest, step, from_dict methods within AppWorldEnv.

Also add detailed annotations about how AppWorldEnv interagte with AppWorld. This commit doesn't include real executions of AppWorld shell. Following code changes will cover the executions of AppWorld shell.

* Implement methods interacting with env within AppWorldReact Agent

This change majorly includes implementations of basic methods: init, reset, and update_from_env of AppWorldReact Agent.

Interactions with model: update_from_model will be implemented in next commit.

* Add AppWorld example to initialize the engine and sample tasks

For this initial version, we will 1) use online open ai models instead of local LLM servers; 2) Load standardized tasks from AppWorld instead of customize AppWorld enrironments

* Implement update_from_model in AppWorld Agent

This commit covers the implementation of the critical part of update_from_model in the agent including:
1) extract the code action from the model response
2) create a new step for agent and append to the trajectory

* Add support of mock mode if AppWorld fails to initialize

In order to test the flow end to end even when the AppWorld is not imported suffessfully. The AppWorld env will simulate a mock mode to fake a response

* Create mock tasks in example if AppWorld cannot be loaded

* Execute the env in real appworld

Spcifically, this commit:
1) Load the World corresponding to the task_id
2) Execute the code in the world and get a 'real' response

* Load officail task ids from AppWorld and apply into AppWorld example

This commit also adds a patch for signal package that's used in AppWorld that prevents rllm to run AppWorld in multi-thread

* Add a global lock to serialize AppWorld operations due to its shared class-level cache

Fix the thread safety issue surfaced in AppWorld when creating multi AppWorld instances since AppWorld has a class level shared cache.

So this commit add a global RLock for the appworld to allow reentrant

* Set max turns

* Add argparse to customize number of tasks etc

* Add accuracy statistics

* Fix evaluation

* remove mock mode

* Use standardized AppWorld react prompt

* Fix evaluation

* fix typo

* Move appworld import ahead

To avoid changes when freezegun starts the sys.modules traversal, move import of appworld to the begining of the file. This can effectively solve the error:

  File "/mnt/space/vamsi/agentic_workflows/rllm/rllm/environments/appworld/appworld_env.py", line 70, in reset
    self.world = AppWorld(task_id=task_id)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/appworld/environment.py", line 541, in __init__
    self.initialize()
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/appworld/environment.py", line 593, in initialize
    self._execute_preamble()
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/appworld/environment.py", line 636, in _execute_preamble
    self.apis, self.requester = ApiCollection.load(
                                ^^^^^^^^^^^^^^^^^^^
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/appworld/collections/apis.py", line 148, in load
    requester = Requester(
                ^^^^^^^^^^
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/appworld/requester.py", line 354, in __init__
    self.setup()
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/appworld/requester.py", line 372, in setup
    self.time_freezer_or_id = set_local_date_and_time(self.date_and_time)
                              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/appworld/apps/api_lib.py", line 644, in set_local_date_and_time
    freezer.start()
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/appworld/common/utils.py", line 4290, in start
    self._freezer.start()
  File "/mnt/remote/mengmengj/conda/envs/rllm311/lib/python3.11/site-packages/freezegun/api.py", line 773, in start
    for mod_name, module in list(sys.modules.items()):
                            ^^^^^^^^^^^^^^^^^^^^^^^^^
RuntimeError: dictionary changed size during iteration

* Extract supervisor information from task and insert to the system prompt template

 Please enter the commit message for your changes. Lines starting

* Split initial prompt into multiple chat messages based on roles

Originally all the messages are passed in as a whole piece. But what AppWorld does is split them into multiple messages

* Add AppWrold React agent test suite

* Remove unnecessary loggings

---------

Co-authored-by: Sunan Xiang <xsunan@cs.uchicago.edu>